Revert "zram: switch to 64-bit hash for dedup"
This reverts commit a749ab6170c78270326e883b85deedaaaacf0889.

+++ drivers/block/zram/zram_drv.c
@@ -1343,7 +1343,7 @@ static int __zram_bvec_write(struct zram *zram, struct bio_vec *bvec,
 	void *src, *dst, *mem;
 	struct zcomp_strm *zstrm;
 	struct page *page = bvec->bv_page;
-	u64 checksum;
+	u32 checksum;
 	unsigned long element = 0;
 	enum zram_pageflags flags = 0;
